+++ clients/dotnet/Microsoft.IoT.ModelsRepository.CommandLine/src/ModelImporter.cs
@@ -3,7 +3,9 @@
 
 using Azure.IoT.ModelsRepository;
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Threading.Tasks;
 
 namespace Microsoft.IoT.ModelsRepository.CommandLine
 {
@@ -12,19 +14,39 @@ namespace Microsoft.IoT.ModelsRepository.CommandLine
         public static void Import(string modelContent, DirectoryInfo repository)
         {
             string rootId = ParsingUtils.GetRootId(modelContent);
-            string createPath = DtmiConventions.GetModelUri(rootId, new Uri(repository.FullName)).AbsolutePath;
+            string createPath = DtmiConventions.GetModelUri(rootId, new Uri(repository.FullName)).LocalPath;
 
-            Outputs.WriteOut($"- Importing model \"{rootId}\"...");
+            Outputs.WriteOut($"* Importing model \"{rootId}\".");
             if (File.Exists(createPath))
             {
                 Outputs.WriteOut(
-                    $"- Skipping \"{rootId}\". Model file already exists in repository.",
+                    $"* Skipping \"{rootId}\". Model file already exists in repository.",
                     ConsoleColor.DarkCyan);
                 return;
             }
 
             (new FileInfo(createPath)).Directory.Create();
             Outputs.WriteToFile(createPath, modelContent);
+        }
+
+        public static async Task<int> ImportFileAsync(FileInfo modelFile, DirectoryInfo repository, RepoProvider repoProvider, ValidationRules rules=null)
+        {
+            int validationResult = await Validations.ValidateModelFileAsync(modelFile, repoProvider, rules);
+            if (validationResult != ReturnCodes.Success)
+            {
+                return validationResult;
+            }
+
+            // TODO: Redundant
+            FileExtractResult extractResult = ParsingUtils.ExtractModels(modelFile);
+            List<string> models = extractResult.Models;
+
+            foreach (string content in models)
+            {
+                Import(content, repository);
+            }
+
+            return ReturnCodes.Success;
         }
     }
 }