+++ clients/dotnet/Microsoft.IoT.ModelsRepository.CommandLine/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.0.0-beta.5 (unreleased)
 
+- The `validate` command supports validation of directory model files (recursive) with file search pattern.
+- The `import` command supports the import of directory model files (recursive) with file search pattern.
+- Removed the `--strict` flag from the `import` command. When importing model files to an IoT Models Repository strict validation
+  is implicit.
 - Upgrade Azure.IoT.ModelsRepository dependency to 1.0.0-preview.4
 - Upgrade System.CommandLine dependency to 2.0.0-beta1.21308.1